backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
python-Levenshtein
python-dateutil
python-dotenv
fuzzywuzzy
google
matplotlib
mysql-connector-python
google-generativeai
numpy
openpyxl
pandas
xlrd
gunicorn
plotly
scipy
python-pptx
seaborn
SQLAlchemy
XlsxWriter
PyJWT
Werkzeug
pymysql
cryptography
